+++ src/rules/parse.cc
@@ -67,12 +67,24 @@ static std::optional<std::string> validate_rule(Rule &rule)
             return "Using a reject action in conjuction with a socket"
                    " path is not allowed.";
 
+        if (rule.ignore)
+            return "Using an ignore action in conjuction with a socket"
+                   " path is not allowed.";
+
         if (rule.blackhole)
             return "Using a blackhole action in conjuction with a socket"
                    " path is not allowed.";
     } else if (rule.reject && rule.blackhole) {
         return "Reject and blackhole actions are mutually exclusive.";
-    } else if (rule.reject) {
+    } else if (rule.ignore && (rule.blackhole || rule.reject)) {
+        return "Ignore action can't be used in conjunction with blackhole"
+               " or reject.";
+#ifdef SOCKET_ACTIVATION
+    } else if (rule.ignore && rule.socket_activation) {
+        return "Ignore action can't be used in conjunction with socket"
+               " activation.";
+#endif
+    } else if (rule.reject || rule.ignore) {
         return std::nullopt;
     } else if (rule.blackhole) {
         if (rule.direction != RuleDir::INCOMING)
@@ -80,11 +92,11 @@ static std::optional<std::string> validate_rule(Rule &rule)
         return std::nullopt;
 #ifdef SOCKET_ACTIVATION
     } else if (!rule.socket_activation) {
-         return "Socket activation is disabled and no socket"
-                " path, reject or blackhole action was specified.";
+        return "Socket activation is disabled and no socket"
+               " path, reject, ignore or blackhole action was specified.";
 #else
     } else {
-        return "No socket path, reject or blackhole action specified.";
+        return "No socket path, reject, ignore or blackhole action specified.";
 #endif
     }
 
@@ -216,6 +228,8 @@ static std::optional<Rule> parse_rule(const std::string &file, int pos,
             }
         } else if (key == "blackhole") {
             RULE_CONVERT(rule.blackhole, "blackhole", bool, "bool");
+        } else if (key == "ignore") {
+            RULE_CONVERT(rule.ignore, "ignore", bool, "bool");
         } else if (key == "socketPath") {
             RULE_CONVERT(rule.socket_path, "socketPath", std::string,
                          "string");
@@ -393,6 +407,8 @@ std::optional<Rule> parse_rule_arg(size_t rulepos, const std::string &arg)
                 rule.reject = true;
             } else if (buf == "blackhole") {
                 rule.blackhole = true;
+            } else if (buf == "ignore") {
+                rule.ignore = true;
             } else {
                 print_arg_error(rulepos, arg, errpos, errlen, "unknown flag");
                 return std::nullopt;
@@ -474,6 +490,9 @@ std::string encode_rules(std::vector<Rule> rules)
         if (rule.blackhole)
             node["blackhole"] = true;
 
+        if (rule.ignore)
+            node["ignore"] = true;
+
         doc.push_back(node);
     }
 
@@ -542,6 +561,8 @@ void print_rules(std::vector<Rule> &rules, std::ostream &out)
                 out << "." << std::endl;
             } else if (rule.blackhole) {
                 out << "  Blackhole the socket." << std::endl;
+            } else if (rule.ignore) {
+                out << "  Don't handle this socket." << std::endl;
             } else {
                 out << "  Socket path: " << rule.socket_path.value()
                     << std::endl;